Feat - user controller errors

+++ src/controllers/user.ts
@@ -2,6 +2,7 @@ import { Request, Response } from 'express'
 import { UserService } from '../services/user'
 import { userCreateSchema } from '../requestSchemas/user'
 import { StatusCodes } from 'http-status-codes'
+import { ApiError } from '../error/ApiError'
 
 export class UserController {
   private _service: UserService
@@ -21,6 +22,7 @@ export class UserController {
   }
 
   public async deleteUser(req: Request, res: Response){
+    if (!Number(req.params.id)) return ApiError.badRequest('Id property type needs to be Integer')
     const deleted = await this._service.deleteUser(Number(req.params.id))
     res.status(StatusCodes.OK).json({ deleted })
   }
